refactor: set up types for login buttons and exec transactions
