--- sdk/src/types/BaseDimo.ts
@@ -1,31 +1,0 @@
-import { EntryState } from "../enums/globalEnums";
-
-export interface DynamicButtonLabels {
-  authenticatedLabel?: string;
-  unAuthenticatedLabel?: string;
-}
-
-export interface ShareBaseDimoButton {
-  mode: "popup" | "embed" | "redirect";
-  onSuccess: (authData: { token: string }) => void; // Success callback
-  onError: (error: Error) => void; // Error callback
-}
-
-export interface BasePayload extends BasePayloadParams {
-  onSuccess: (data: {
-    token: string;
-    transactionHash?: string;
-    transactionReceipt?: any;
-  }) => void;
-  onError: (error: Error) => void;
-  setAuthenticated: React.Dispatch<React.SetStateAction<boolean>>;
-}
-
-export interface BasePayloadParams {
-  entryState: EntryState;
-  dimoLogin: string;
-  forceEmail: boolean;
-  clientId?: string;
-  redirectUri?: string;
-  apiKey?: string; // Avoid sending API key in the URL
-}
